requirements.txt
streamlit
beautifulsoup4
requests
langchain
langchain-community
langchain-huggingface
pinecone-client
sentence-transformers
pydantic
python-dotenv